test: Move and restructure HybridKeyExchange test
