+++ SSH-Core/src/main/java/de/rub/nds/sshattacker/core/crypto/kex/HybridKeyExchange.java
@@ -158,7 +158,7 @@ public abstract class HybridKeyExchange extends KeyExchange {
                             + ArrayConverter.bytesToRawHexString(
                                     encode(tmpSharedSecret, algorithm.getDigest())));
         } catch (Exception e) {
-            LOGGER.warn("Could not create the shared Secret: " + e);
+            LOGGER.warn("Could not create the shared secret: " + e);
         }
     }
